requirements.txt
httpx
python-dotenv
beautifulsoup4
socksio